refactor: Utilize ProductColorValueFormAdmin in ProductColorValueTabu
This commit refactors the `ProductColorValueTabu` inline form in the store app to use the `ProductColorValueFormAdmin` form. By specifying the form class, the admin interface benefits from enhanced inventory validation and improved management capabilities when dealing with color values associated with products.

refactor: Improve inventory management in ProductAdmin

This commit refactors the `ProductAdmin` in the store app to enhance inventory management. The `save_model` method is reset the adjusted value in the form instance or object to match what the user entered. Additionally, the `get_formsets_with_inlines` method is updated to pass the product inventory to the `ProductColorValueTabu` inline form for improved validation and consistency when dealing with color values associated with products. This ensures a more streamlined and reliable process for handling inventory changes within the admin interface.