feat: Add ProductColorAndSizeValueTabu admin.TabularInline
This commit introduces the ProductColorAndSizeValueTabu admin.TabularInline to the Product admin. The inline is designed for managing color and size values associated with a product. It includes form customization and presents relevant fields for color, size, size price, inventory, and color image. Additionally, the inline provides a method for displaying color images and optimizes database queries using select_related.

feat: Add ProductImageTabu admin.TabularInline

This commit adds the ProductImageTabu admin.TabularInline to the Product admin. The inline facilitates the management of product images, including fields for image and is_main. The form associated with this inline includes validation to ensure that only one image is marked as the main image.

feat: Add ProductCommentTabu admin.TabularInline

This commit introduces the ProductCommentTabu admin.TabularInline to the Product admin. The inline is designed for handling product comments, providing fields for author, text, confirmation, and datetime_updated. The inline includes autocomplete fields for the author and formfield_overrides for a better user interface.